requirements.txt
# Core deep learning and machine learning libraries
tensorflow==2.16.1
keras==2.16.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Computer vision and image processing
opencv-python==4.8.1.78
Pillow==10.0.0
matplotlib==3.7.2
seaborn==0.12.2

# Web framework for inference
streamlit==1.28.1
flask==2.3.3

# Additional utilities
tqdm==4.66.1
jupyter==1.0.0
ipykernel==6.25.2

# Development and testing
pytest==7.4.2
